requirements.txt
# Web Framework
Flask>=2.3.0,<4.0.0
flask-sqlalchemy>=3.0.0,<4.0.0
flask-migrate>=4.0.0,<5.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0

# HTTP Requests
requests>=2.28.0,<3.0.0

# Environment Variables
python-dotenv>=0.19.0,<2.0.0

# AI/LLM Integration
langchain-google-genai>=0.1.0,<2.0.0
langchain-core>=0.1.0,<1.0.0

# Retry Logic
tenacity>=8.0.0,<9.0.0

# Cloud Storage
cloudinary>=1.30.0,<2.0.0

# Development and Debugging (optional)
# Uncomment if needed for development
# pytest==8.3.2
# pytest-flask==1.3.0
# black==24.8.0
# flake8==7.1.1

# Production Server (optional)
# Uncomment for production deployment
# gunicorn==23.0.0
# uwsgi==2.0.26
